feat: create pgsql-client package with DbAdmin and PgClient utilities
Extract reusable PostgreSQL client utilities from pgsql-test into a new
pgsql-client package:

- DbAdmin: Database administration (create, drop, templates, extensions, grants)
- PgClient: Query helpers (any, one, oneOrNone, many, etc.) with RLS context
- Role utilities: Role name mapping for anonymous, authenticated, administrator
- Context utilities: Generate SQL for setting PostgreSQL session context
- Stream utilities: Stream SQL to psql process

pgsql-test now extends pgsql-client:
- PgTestClient extends PgClient, adding test lifecycle methods (beforeEach, afterEach, publish)
- DbAdmin extends base DbAdmin, adding test-specific createSeededTemplate method
- Other utilities re-exported from pgsql-client for backward compatibility

This separation allows DbAdmin and PgClient to be used in scripts, CLI tools,
and applications without pulling in test-specific dependencies.
