+++ postgres/pgsql-test/package.json
@@ -59,15 +59,15 @@
     "@types/pg-copy-streams": "^1.2.5",
     "makage": "^0.1.9"
   },
-  "dependencies": {
-    "@pgpmjs/core": "workspace:^",
-    "@pgpmjs/env": "workspace:^",
-    "@pgpmjs/logger": "workspace:^",
-    "@pgpmjs/server-utils": "workspace:^",
-    "@pgpmjs/types": "workspace:^",
-    "pg": "^8.16.3",
-    "pg-cache": "workspace:^",
-    "pg-env": "workspace:^",
-    "pgsql-seed": "workspace:^"
-  }
+    "dependencies": {
+      "@pgpmjs/env": "workspace:^",
+      "@pgpmjs/logger": "workspace:^",
+      "@pgpmjs/server-utils": "workspace:^",
+      "@pgpmjs/types": "workspace:^",
+      "pg": "^8.16.3",
+      "pg-cache": "workspace:^",
+      "pg-env": "workspace:^",
+      "pgsql-client": "workspace:^",
+      "pgsql-seed": "workspace:^"
+    }
 }
