+++ postgres/pgsql-test/src/index.ts
@@ -5,3 +5,6 @@ export * from './roles';
 export * from './seed';
 export * from './test-client';
 export { snapshot } from './utils';
+
+// Re-export PgClient and PgClientOpts from pgsql-client for convenience
+export { PgClient, PgClientOpts, generateContextStatements, streamSql } from 'pgsql-client';
